[COMPAT] Move Puzzles::set to compat::set
