+++ tests/maths/maths_primes_test.cpp
@@ -22,15 +22,15 @@
 
 #include "maths/primes.h"
 
-#include "common/set.h"
+#include "compat/set.h"
 
 #include <gtest/gtest.h>
 
 using namespace Maths;
 
 TEST(Maths, IsPrime) {
-  Puzzles::set<uint_fast8_t> primes = {2,  3,  5,  7,  11, 13, 17, 19, 23, 29, 31, 37, 41,
-                                       43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97};
+  compat::set<uint_fast8_t> primes = {2,  3,  5,  7,  11, 13, 17, 19, 23, 29, 31, 37, 41,
+                                      43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97};
   for (auto i = 1; i <= 100; ++i) {
     if (primes.contains(i)) {
       EXPECT_TRUE(isPrime(i)) << "Prime number " << i << " was reported as NOT being prime";
